Remove code duplication in runners
